+++ bindings/nodejs/lib/types/block/payload/transaction/index.ts
@@ -2,5 +2,5 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export * from './essence';
-export * from './transaction';
 export * from './unlock';
+export * from './transaction';